requirements.txt
PyPDF2
PyMuPDF
python-pptx
python-docx
pytesseract
Pillow
fitz
fastapi
uvicorn
python-multipart